fix: fix scrollbar of tooltip
