+++ packages/vchart/src/plugin/components/tooltip-handler/constants.ts
@@ -1,7 +1,9 @@
+import type { ITooltipHandlerOptions } from './interface';
+
 export const TOOLTIP_CONTAINER_EL_CLASS_NAME = 'vchart-tooltip-container';
 export const TOOLTIP_EMPTY_STRING = '';
 
-export const DEFAULT_OPTIONS = {
+export const DEFAULT_OPTIONS: ITooltipHandlerOptions = {
   /**
    * X offset.
    */
@@ -12,6 +14,3 @@ export const DEFAULT_OPTIONS = {
    */
   offsetY: 10
 };
-
-// FIXME: 命名规范
-export type Options = typeof DEFAULT_OPTIONS;
