--- packages/vchart/src/plugin/components/tooltip-handler/dom/model/interface.ts
@@ -1,11 +1,0 @@
-import type { ITooltipActual, Maybe } from '../../../../../typings';
-import type { IDomTooltipStyle } from '../interface';
-import type { ITooltipAttributes } from '../../interface';
-
-export interface ITooltipModelOption {
-  valueToHtml: (value: any) => string;
-  getTooltipStyle: () => IDomTooltipStyle;
-  getTooltipActual: () => Maybe<ITooltipActual>;
-  getTooltipAttributes: () => Maybe<ITooltipAttributes>;
-  getContainer: () => HTMLElement;
-}
