--- packages/vchart/src/plugin/components/tooltip-handler/interface/common.ts
@@ -1,5 +1,0 @@
-import type { ITooltipHandler } from '../../../../typings';
-
-export interface ITooltipHandlerConstructor {
-  new (): ITooltipHandler;
-}
